style: LinkedList 파일 생성 및 코드 분리
